Extract FrameRate to a separate module to break an import cycle
